+++ src/components/index.ts
@@ -3,5 +3,4 @@ import { createElement } from "react"
 
 setup(createElement);
 
-export { BreakpointerIndicator } from "./breakpointer.indicator"
 export { BreakpointerProvider } from "./breakpointer.provider"