+++ src/contexts/breakPoint.context.tsx
@@ -3,8 +3,6 @@ import { createContext } from "react"
 interface IBreakPointState {
     currentWidth: number,
     currentScreen: string | null,
-    mode: string
-
 }
 
 interface IBreakPointContext extends IBreakPointState {
@@ -12,8 +10,7 @@ interface IBreakPointContext extends IBreakPointState {
 
 const BreakpointerContext = createContext<IBreakPointContext>({
     currentWidth: window.innerWidth,
-    currentScreen: "",
-    mode: "development"
+    currentScreen: ""
 })
 
 export default BreakpointerContext