+++ src/core/unit.parser.ts
@@ -1,11 +1,18 @@
-export const parseUnit = (obj: Record<string, string>) => {
+export const parseUnit = (obj: Record<string, string | number>) => {
     let parsedObj: Record<string, number> = {};
 
     for (let key in obj) {
-        parsedObj[key] = parseInt(obj[key]);
+        switch (typeof obj[key]) {
+            case "number":
+                parsedObj[key] = obj[key];
+                break;
+            case "string":
+                parsedObj[key] = parseInt(obj[key]);
+                break;
+        }
     }
 
     return parsedObj
 }
 
-// WRITE Tests
+// TODO - test cases